Add files via upload
This is the week 1 assessment of the course.  In this exercise, we wrote code to predict house prices based on a simple linear equation. The x and y input variables are just numpy arrays created by us. Then we created a single Dense Layer with 1 unit. Then we fit the model with x and y variables and run for 1000 epochs. We finally predicted the output variable for a given number.
This assignment received a 100 % grade.